Split add task and display info modals into separate files
